Split TestCoreTaskExtensions into separate files for Finally, Select, and Then tests
