+++ Tests/UnitTest.RackspaceThreading/UnitTest.RackspaceThreading.portable-net40.csproj
@@ -61,7 +61,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestCancellationTokenSourceExtensions.cs" />
     <Compile Include="TestCompletedTask.cs" />
-    <Compile Include="TestCoreTaskExtensions.cs" />
+    <Compile Include="TestCoreTaskExtensions_Finally.cs" />
+    <Compile Include="TestCoreTaskExtensions_Select.cs" />
+    <Compile Include="TestCoreTaskExtensions_Then.cs" />
     <Compile Include="TestProgress.cs" />
     <Compile Include="TestTaskBlocks_Using.cs" />
     <Compile Include="TestTaskBlocks_While.cs" />
